Add requirements.txt for required python packages
The file requirements.txt contains a list of all python packages that
need to be installed via 'pip install'.

With this file it's easier to know which packages have to be installed.
The installation process can even be made easier as the file can be
passed as a command line option to pip.