Add wrapper class for connection to car
To make it possible to use one serial connection to both send data to
and receive data from the car, a wrapper class for serial communication
and the AMB8826 library is added.

The CarConnection class provides one serial connection and methods to
send and receive data, which call the corresponding functions in the
AMB8826 library.