Rewrite webserver for websocket connections
Add possibility for handling multiple websocket connections.
Change behaviour of receiving data to be completely asynchronous.

That means, that receiving data from the other process through a pipe
and sending this data to all clients can be done in one thread.
This is possible because asynchronous functions can give up control to
other functions. When their data is ready the asynchronous functions can
take back control and continue working.