Fix bug resulting from change in conversion to hex
In v2.0.0 the conversion from bytes to hex was changed in order to fix a
bug with a leading 0 in a byte. This resulted in the conversion
returning '00' instead of '0'. This change resulted in the end of the
package not being recognized correctly anymore.

The result of the conversion from byte to hex now gets converted into an
int and is then compared to 0. In this way '00' and '0' both result in
the end of the package being recognized correctly.