Alter configuration and optimize data processing
Change the way the server is configured to be more userfriendly.
Optimize the processing of data to be more efficient.

Before there have been two configuration files that are now combined in
a better structured file written in TOML. The data structure into which
the configuration is loaded is changed also to be more logical. A class
which wraps the configuration is added.

The processing of data is not wrapped in a class anymore, as no state is
needed in this part of the code (except configuration, which is passed
as argument). Unnecessary parts of the code are cleaned up.